requirements.txt
# Database
mysql-connector-python==8.2.0
python-dotenv==1.0.0

# Data Processing
pandas==2.1.4
numpy==1.26.2

# Visualization
matplotlib==3.8.2
seaborn==0.13.0
